Update README and set up functional Lambda deploys
